Move segment package for same level with statement (#10846)
