+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/test/java/org/apache/shardingsphere/proxy/backend/text/distsql/rdl/impl/CreateShardingBindingTableRulesBackendHandlerTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl;
 
-import org.apache.shardingsphere.sharding.distsql.parser.statement.segment.ShardingBindingTableRuleSegment;
+import org.apache.shardingsphere.sharding.distsql.parser.segment.BindingTableRuleSegment;
 import org.apache.shardingsphere.sharding.distsql.parser.statement.CreateShardingBindingTableRulesStatement;
 import org.apache.shardingsphere.infra.context.metadata.MetaDataContexts;
 import org.apache.shardingsphere.infra.metadata.ShardingSphereMetaData;
@@ -127,17 +127,17 @@ public final class CreateShardingBindingTableRulesBackendHandlerTest {
     }
     
     private CreateShardingBindingTableRulesStatement buildShardingTableRuleStatement() {
-        ShardingBindingTableRuleSegment segment = new ShardingBindingTableRuleSegment();
+        BindingTableRuleSegment segment = new BindingTableRuleSegment();
         segment.setTables("t_order,t_order_item");
-        ShardingBindingTableRuleSegment segmentAnother = new ShardingBindingTableRuleSegment();
+        BindingTableRuleSegment segmentAnother = new BindingTableRuleSegment();
         segmentAnother.setTables("t_1,t_2");
         return new CreateShardingBindingTableRulesStatement(Arrays.asList(segment, segmentAnother));
     }
     
     private CreateShardingBindingTableRulesStatement buildDuplicateShardingTableRuleStatement() {
-        ShardingBindingTableRuleSegment segment = new ShardingBindingTableRuleSegment();
+        BindingTableRuleSegment segment = new BindingTableRuleSegment();
         segment.setTables("t_order,t_order_item");
-        ShardingBindingTableRuleSegment segmentAnother = new ShardingBindingTableRuleSegment();
+        BindingTableRuleSegment segmentAnother = new BindingTableRuleSegment();
         segmentAnother.setTables("t_order,t_order_item");
         return new CreateShardingBindingTableRulesStatement(Arrays.asList(segment, segmentAnother));
     }
